Health Bar Implementation

+++ Assets/HealthBar.cs
@@ -10,18 +10,15 @@ public class HealthBar : MonoBehaviour
     // variable to store the slider
     public Slider slider;
 
-    public void setMaxHealth(int health)
+    public void setMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health; // sliders should start at max health 
 
     }
 
-    public void setHealth(int health)
+    public void setHealth(float health)
     {
         slider.value = health;
     }
-
-
-
 }
